# docs/swagger.yaml
basePath: /
definitions:
  account.GetAccountVO:
    description: 请求获取账户信息时所需参数
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
    type: object
  account.LoginVO:
    description: 登录成功后返回的访问令牌和刷新令牌
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  account.UpdateAccountVO:
    description: 更新账户信息后返回的参数
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
    type: object
...
  dto.LoginRequest:
    description: 用户登录请求所需参数
    properties:
      email:
        type: string
      img_verification_code:
        type: string
      password:
        type: string
    required:
    - email
    - img_verification_code
    - password
    type: object
  dto.RegisterRequest:
    description: 用户注册所需参数
    properties:
      email:
        type: string
      email_verification_code:
        type: string
      img_verification_code:
        type: string
      nickname:
        maxLength: 20
        minLength: 1
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - email
    - email_verification_code
    - img_verification_code
    - nickname
    - password
    type: object
  dto.ResetPwdRequest:
    description: 用户重置密码所需参数
    properties:
      again_new_password:
        maxLength: 20
        minLength: 6
        type: string
      email:
        type: string
      email_verification_code:
        type: string
      new_password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - again_new_password
    - email
    - email_verification_code
    - new_password
    type: object
  dto.UpdateAccountRequest:
    description: 用户更新账户信息所需参数
    properties:
      avatar:
        type: string
      nickname:
        maxLength: 20
        minLength: 1
        type: string
      phone:
        type: string
    required:
    - nickname
    type: object
...
  vo.Result:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 响应数据
      msg:
        description: 错误信息
        type: string
      requestId:
        description: 请求ID
      timeStamp:
        description: 响应时间戳
    type: object
host: localhost:9010
info:
  contact: {}
  description: This is the API documentation for Jank Blog.
  version: "1.0"
paths:
  /account/getAccount:
    get:
      consumes:
      - application/json
      description: 根据提供的邮箱获取对应用户的详细信息
      parameters:
      - in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/account.GetAccountVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取账户信息
      tags:
      - 账户
  /account/loginAccount:
    post:
      consumes:
      - application/json
      description: 用户登录并获取访问令牌，支持图形验证码校验
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      - description: 图形验证码
        in: query
        name: ImgVerificationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回访问令牌
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/account.LoginVO'
              type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: '#/definitions/vo.Result'
        "401":
          description: 登录失败，凭证无效
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 用户登录
      tags:
      - 账户
  /account/logoutAccount:
    post:
      description: 退出当前用户登录状态
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 账户
  /account/registerAccount:
    post:
      consumes:
      - application/json
      description: 注册新用户账号，支持图形验证码和邮箱验证码校验
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      - description: 图形验证码
        in: query
        name: ImgVerificationCode
        required: true
        type: string
      - description: 邮箱验证码
        in: query
        name: EmailVerificationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterRequest'
              type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 用户注册
      tags:
      - 账户
  /account/resetPassword:
    post:
      consumes:
      - application/json
      description: 重置用户账户密码，支持邮箱验证码校验
      parameters:
      - description: 重置密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPwdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: '#/definitions/vo.Result'
        "401":
          description: 未授权，用户未登录
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 账户
  /account/updateAccount:
    post:
      consumes:
      - application/json
      description: 更新当前登录用户的账户信息
      parameters:
      - description: 更新账户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/account.UpdateAccountVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 更新账户信息
      tags:
      - 账户
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"